refactor holidays and irish bingo to separate server side page and client components
