+++ src/components/backtotop.tsx
@@ -1,3 +1,5 @@
+"use client";
+
 import { useWindowScroll } from "@mantine/hooks";
 import { Button, Transition, rem, Affix } from "@mantine/core";
 import { IconArrowUp } from "@tabler/icons-react";
